influx: Use sync2.Sleep in flush method
Use the sync2.Sleep function rather than the standard time.Sleep for
considering context cancellations.

This commit doesn't change the method signature, so it isn't possible to
cancel through the used context, however, it prepares the method for a
future refactoring where the context is passed to method to not forget
to apply these changes at that time.

Change-Id: I49c05cc76114555fa3a364c35c7c4d13f6e875c1